+++ app/uk/gov/hmrc/customs/declarations/information/action/DeclarationFullCheckAction.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.customs.declarations.information.controllers.actionBuilders
+package uk.gov.hmrc.customs.declarations.information.action
 
 import play.api.http.HttpEntity
 import play.api.mvc.{ActionRefiner, Result}
