+++ app/uk/gov/hmrc/customs/declarations/information/action/InternalClientIdsCheckAction.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.customs.declarations.information.controllers.actionBuilders
+package uk.gov.hmrc.customs.declarations.information.action
 
 import play.api.mvc.{ActionRefiner, Result}
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse.errorBadRequest
