+++ app/uk/gov/hmrc/customs/declarations/information/action/ShutterCheckAction.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.customs.declarations.information.controllers.actionBuilders;
+package uk.gov.hmrc.customs.declarations.information.action
 
 import play.api.http.HeaderNames.ACCEPT
 import play.api.http.Status.SERVICE_UNAVAILABLE
